Combined index,forever,bot into a new index.js
